requirements.txt
botogram2==0.6.1
certifi==2020.11.8
chardet==3.0.4
idna==2.10
Logbook==1.5.3
plumbum==1.6.9
PyYAML==5.3.1
requests==2.25.0
urllib3==1.26.2
